Release v1.2.9: fix auth avatar loading, right-aligned chat layout, join-page UI, and ESLint warnings.
