+++ client/package.json
@@ -21,6 +21,7 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
+  "proxy": "http://localhost:7777",
   "eslintConfig": {
     "extends": [
       "react-app",
