--- templates/index.html
@@ -1,67 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Real-Time Chat</title>
-    <style>
-        body { font-family: Arial, sans-serif; }
-        #messages { list-style-type: none; padding: 0; }
-        #messages li { padding: 5px 10px; }
-        #form { display: flex; }
-        #input { flex-grow: 1; }
-    </style>
-</head>
-<body>
-    <ul id="messages"></ul>
-    <form id="form" action="">
-        <input id="input" autocomplete="off" /><button>Send</button>
-    </form>
-    <script src="/socket.io/socket.io.js"></script>
-    <script>
-        var socket = io();
-        var form = document.getElementById('form');
-        var input = document.getElementById('input');
-        var messages = document.getElementById('messages');
-        var username = prompt("Enter your username:");
-        var room = '123'; // The room we want to join, as you mentioned
-    
-        // Set username when user connects
-        socket.emit('set username', username);
-
-        // Join the specific chat room
-        joinRoom(room);
-    
-        // Add listener for chat history upon joining a room
-        socket.on('chat history', function(history) {
-            messages.innerHTML = ''; // Clear out any existing messages
-            history.forEach(function(msg) {
-                var item = document.createElement('li');
-                item.textContent = msg.username + ': ' + msg.text;
-                messages.appendChild(item);
-            });
-            window.scrollTo(0, document.body.scrollHeight);
-        });
-
-        // Add listener for new chat messages
-        socket.on('chat message', function(msg) {
-            var item = document.createElement('li');
-            item.textContent = msg.username + ': ' + msg.text;
-            messages.appendChild(item);
-            window.scrollTo(0, document.body.scrollHeight);
-        });
-    
-        // Handle sending messages
-        form.addEventListener('submit', function(e) {
-            e.preventDefault();
-            if (input.value) {
-                socket.emit('chat message', input.value, room);
-                input.value = '';
-            }
-        });
-    
-        // Allow joining rooms
-        function joinRoom(room) {
-            socket.emit('join room', room);
-        }
-    </script>
-</body>
-</html>
